+++ example/app/models.py
@@ -1,6 +1,12 @@
 from panther.db.models import BaseUser
 
 
-class User(BaseUser):
+class CustomQuery:
+    @classmethod
+    def find_last(cls):
+        return cls.last()
+
+
+class User(BaseUser, CustomQuery):
     username: str
     password: str