+++ example/app/urls.py
@@ -38,6 +38,5 @@ urls = {
     'bg-tasks/': run_background_tasks_api,
     'custom-response/': custom_response_class_api,
     'image/': ImageAPI,
-    'login/': login_api,
     'logout/': logout_api,
 }