+++ example/core/permissions.py
@@ -4,7 +4,7 @@ from panther.request import Request
 
 class UserPermission(BasePermission):
     @classmethod
-    def authorization(cls, request: Request) -> bool:
+    async def authorization(cls, request: Request) -> bool:
         if request.user.username == 'Ali':
             return True
         return False