+++ tests/test_caching.py
@@ -1,6 +1,6 @@
 import time
 from datetime import timedelta
-from unittest import TestCase
+from unittest import IsolatedAsyncioTestCase
 
 from panther import Panther
 from panther.app import API
@@ -9,19 +9,19 @@ from tests._utils import check_two_dicts
 
 
 @API()
-def without_cache_api():
+async def without_cache_api():
     time.sleep(0.01)
     return {'detail': time.time()}
 
 
 @API(cache=True)
-def with_cache_api():
+async def with_cache_api():
     time.sleep(0.01)
     return {'detail': time.time()}
 
 
 @API(cache=True, cache_exp_time=timedelta(seconds=5))
-def expired_cache_api():
+async def expired_cache_api():
     time.sleep(0.01)
     return {'detail': time.time()}
 
@@ -33,34 +33,34 @@ urls = {
 }
 
 
-class TestCaching(TestCase):
+class TestCaching(IsolatedAsyncioTestCase):
     @classmethod
     def setUpClass(cls) -> None:
         app = Panther(__name__, configs=__name__, urls=urls)
         cls.client = APIClient(app=app)
 
-    def test_without_cache(self):
-        res1 = self.client.get('without-cache')
+    async def test_without_cache(self):
+        res1 = await self.client.get('without-cache')
         assert res1.status_code == 200
 
-        res2 = self.client.get('without-cache')
+        res2 = await self.client.get('without-cache')
         assert res2.status_code == 200
 
         assert check_two_dicts(res1.data, res2.data) is False
 
-    def test_with_cache(self):
-        res1 = self.client.get('with-cache')
+    async def test_with_cache(self):
+        res1 = await self.client.get('with-cache')
         assert res1.status_code == 200
 
-        res2 = self.client.get('with-cache')
+        res2 = await self.client.get('with-cache')
         assert res2.status_code == 200
 
         assert check_two_dicts(res1.data, res2.data) is True
 
-    def test_with_cache_5second_exp_time(self):
+    async def test_with_cache_5second_exp_time(self):
         # First Request
         with self.assertLogs(level='INFO') as captured:
-            res1 = self.client.get('with-expired-cache')
+            res1 = await self.client.get('with-expired-cache')
         assert res1.status_code == 200
 
         # Check Logs
@@ -68,7 +68,7 @@ class TestCaching(TestCase):
         assert captured.records[0].getMessage() == '"cache_exp_time" is not very accurate when redis is not connected.'
 
         # Second Request
-        res2 = self.client.get('with-expired-cache')
+        res2 = await self.client.get('with-expired-cache')
         assert res2.status_code == 200
 
         # Response should be cached
@@ -77,7 +77,7 @@ class TestCaching(TestCase):
         time.sleep(5)
 
         # Third Request
-        res3 = self.client.get('with-expired-cache')
+        res3 = await self.client.get('with-expired-cache')
         assert res3.status_code == 200
 
         # After 5 seconds we should have a new response